todo/requirements.txt
Flask==2.2.2
requests==2.28.2
opentelemetry-api==1.15.0
opentelemetry-exporter-jaeger==1.15.0
opentelemetry-exporter-jaeger-thrift==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-instrumentation==0.36b0   
